pyproject.toml
[project]
name = "google-ads-mcp-complete"
version = "2.0.0"
description = "Complete Google Ads API v21 MCP Server - All 29 tools fully implemented and working"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0,<2",
    "google-ads>=29.0.0",
    "pydantic>=2.0.0,<3",
    "httpx>=0.27.0,<1",
    "tenacity>=8.0.0,<10",
    "cachetools>=5.0.0,<6",
    "structlog>=24.0.0,<26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
    "mypy>=1.0.0"
]

[build-system]
requires = ["setuptools>=70.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true